Move AppConfig into own module dir
