+++ apps/typeform/frontend/src/AppConfig/index.tsx
@@ -1,0 +1,1 @@
+export { AppConfig } from './AppConfig';
