+++ apps/typeform/frontend/src/AppConfig/styles.ts
