+++ apps/typeform/frontend/src/typings.ts
